Extract utf8ToUtf16String() to generic fsnotifier
